examples/twooter/requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.5
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.4
WTForms==1.0.5
Werkzeug==0.9.6
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.2
itsdangerous==0.24
psycopg2==2.5.3
pyzmq==14.3.1
wsgiref==0.1.2
